publishers: add slug to publisher URLs
Temporarily move SlugDetailView under publishers/
