+++ dissemin/urls.py
@@ -92,7 +92,7 @@ urlpatterns = patterns('',
     # Apps
     url(r'^ajax-upload/', include('upload.urls')),
     url(r'^', include('papers.urls')),
-    url(r'^publisher', include('publishers.urls')),
+    url(r'^', include('publishers.urls')),
     url(r'^', include('deposit.urls')),
     url(r'^', include('notification.urls')),
     url(r'^jsreverse/$', 'django_js_reverse.views.urls_js', name='js_reverse'),
